qa-app/pyproject.toml
[tool.poetry]
name = "qa-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
langchain = "0.0.333"
fastapi = "0.104.1"
uvicorn = "0.23.2"
python-multipart = "0.0.6"
PyYAML = "6.0.1"
transformers = "4.34.0"
torch = "2.0.1"
tqdm = "4.66.1"
faiss-cpu = "1.7.4"
sentence-transformers = "2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
